+++ backloggd-plus.meta.js
@@ -10,7 +10,7 @@
 // @name:ko           Backloggd Plus
 // @name:pl           Backloggd Plus
 // @namespace         https://github.com/NemoKing1210/backloggd-plus
-// @version           0.5.13
+// @version           0.5.15
 // @description       Extends Backloggd and adds a Backloggd button on Steam game pages
 // @description:ru    Расширяет Backloggd и добавляет кнопку Backloggd на страницах игр Steam
 // @description:zh-CN 扩展 Backloggd：更多游戏信息、更丰富的界面与使用体验
@@ -43,7 +43,6 @@
 // @connect            store.steampowered.com
 // @connect            api.steampowered.com
 // @connect            gamestatus.info
-// @connect            steamdb.info
 // @run-at             document-idle
 // @noframes
 // ==/UserScript==